Update to match refactored metagraph
